currently setting up swagger and application.properties.
