+++ java-backend/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Fri Apr 26 08:58:05 EDT 2019
-gradle.version=5.4.1
+#Fri May 03 15:19:16 EDT 2019
+gradle.version=4.10.3
