+++ java-backend/src/main/resources/static/application.properties
@@ -1,0 +1,11 @@
+#DATABASE CONFIG
+spring.datasource.driver.driverClassName=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost:${port}/postgres
+spring.datasource.username=${username}
+spring.datasource.password=${password}
+
+#EMBEDDED SERVER CONFIG
+#server.address=192.168.0.4
+#bind-address=127.0.0.1
+server.port=8080
+
